Реализует модули
